Adding combobox to react-form
