+++ packages/react-table-ui/src/filters/array-includes-some.tsx
@@ -1,52 +1,7 @@
-import { type HeaderContext, type Row } from "@tanstack/react-table";
-
-import { createCn, Input } from "@zoos/shadcn";
-import { CheckboxGroupVirtual, useVirtualCombobox } from "@zoos/react-form";
 import React from "react";
+import { type HeaderContext } from "@tanstack/react-table";
 
-const Combobox = ({
-  options,
-  inputClassName,
-  ...props
-}: Omit<React.ComponentProps<typeof CheckboxGroupVirtual>, "virtualizer"> & {
-  inputClassName?: string;
-}) => {
-  const {
-    query,
-    setQuery,
-    optionsFiltered,
-    hasUniqueValues,
-    virtualizer,
-    scrollRef,
-  } = useVirtualCombobox({
-    options,
-    virtualizerOptions: { estimateSize: () => 24 },
-  });
-
-  if (!hasUniqueValues) {
-    window.alert(
-      "Options must have unique values. That's one benefit of `Options` as an object with IDs.",
-    );
-  }
-
-  return (
-    <div className="space-y-3">
-      <Input
-        className={inputClassName}
-        value={query}
-        onChange={(e) => setQuery(e.target.value)}
-      />
-      <div className="h-full overflow-auto" ref={scrollRef}>
-        <CheckboxGroupVirtual
-          options={optionsFiltered}
-          virtualizer={virtualizer}
-          className={createCn("flex flex-col gap-3")}
-          {...props}
-        />
-      </div>
-    </div>
-  );
-};
+import { Combobox } from "@zoos/react-form";
 
 const FilterInput = <TData, TValue>({
   headerContext: { column },
